backend/requirements.txt
Django~=5.1.12
djangorestframework~=3.14.0